txn: refactor read pool
